+++ lib/logic/utils/route_generator.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+import 'package:news_aggregator/presentation/pages/home_page.dart';
+
+class RouteGenerator {
+  static Route<dynamic> generateRoute(RouteSettings settings) {
+    switch (settings.name) {
+      case '/':
+        return MaterialPageRoute(
+          builder: ((context) => const HomePage()),
+        );
+      default:
+        return _errorRoute();
+    }
+  }
+
+  static Route<dynamic> _errorRoute() {
+    return MaterialPageRoute(
+      builder: (context) {
+        return const Scaffold(
+          body: Center(
+            child: Text('Page not found!'),
+          ),
+        );
+      },
+    );
+  }
+}
